chore(api): rename main.py to api.py
