+++ da-project-backend/app/__init__.py
@@ -1,1 +1,1 @@
-from .main import app
+from .api import app
